doc-manager-v4/requirements.txt
# Core Logic
watchdog
uuid6
python-dotenv

# Database
psycopg2-binary
pgvector

# AI/ML
openai  # For OpenAI-compatible API (llama.cpp/ollama)

# Monitoring UI
streamlit
pandas
sentence-transformers
